# frontend/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Modular-monolith manufacturing ERP — double-entry ledger, inventory, procure-to-pay, order-to-cash, manufacturing and reporting.",
    "title" : "Manufacturing ERP API",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/purchasing/purchase-orders" : {
      "get" : {
        "operationId" : "listOrders_1",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PurchaseOrderResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "purchasing-controller" ]
      },
      "post" : {
        "operationId" : "createOrder_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurchaseOrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "purchasing-controller" ]
      }
    },
    "/api/purchasing/purchase-orders/{id}" : {
      "get" : {
        "operationId" : "getOrder_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurchaseOrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "purchasing-controller" ]
      }
    },
    "/api/purchasing/purchase-orders/{id}/confirm" : {
      "post" : {
        "operationId" : "confirm_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurchaseOrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "purchasing-controller" ]
      }
    },
...
    "/api/sales/ar-aging" : {
      "get" : {
        "operationId" : "arAging",
        "parameters" : [ {
          "in" : "query",
          "name" : "asOf",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArAgingReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      }
    },
    "/api/sales/customer-returns" : {
      "get" : {
        "operationId" : "listReturns",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CustomerReturnResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      },
      "post" : {
        "operationId" : "postReturn",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateReturnRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerReturnResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      }
    },
    "/api/sales/customer-returns/{id}" : {
      "get" : {
        "operationId" : "getReturn",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerReturnResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      }
    },
...
    "/api/sales/sales-invoices" : {
      "get" : {
        "operationId" : "listInvoices",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SalesInvoiceResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      },
      "post" : {
        "operationId" : "postInvoice",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateInvoiceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesInvoiceResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      }
    },
    "/api/sales/sales-invoices/{id}" : {
      "get" : {
        "operationId" : "getInvoice",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesInvoiceResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      }
    },
    "/api/sales/sales-orders" : {
      "get" : {
        "operationId" : "listOrders",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SalesOrderResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      },
      "post" : {
        "operationId" : "createOrder",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesOrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      }
    },
    "/api/sales/sales-orders/{id}" : {
      "get" : {
        "operationId" : "getOrder",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesOrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      }
    },
    "/api/sales/sales-orders/{id}/confirm" : {
      "post" : {
        "operationId" : "confirm",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesOrderResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "sales-controller" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ArAgingReport" : {
        "type" : "object",
        "properties" : {
          "current" : {
            "type" : "string"
          },
          "days1to30" : {
            "type" : "string"
          },
          "days31to60" : {
            "type" : "string"
          },
          "days61to90" : {
            "type" : "string"
          },
          "days90plus" : {
            "type" : "string"
          },
          "total" : {
            "type" : "string"
          }
        }
      },
...
      "CreateInvoiceRequest" : {
        "type" : "object",
        "properties" : {
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/InvoiceLine"
            },
            "minItems" : 1
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "salesOrderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "taxRateCode" : {
            "type" : "string"
          }
        },
        "required" : [ "lines", "salesOrderId" ]
      },
...
      "CreatePoLine" : {
        "type" : "object",
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "qtyOrdered" : {
            "type" : "string"
          },
          "unitPrice" : {
            "type" : "string"
          }
        },
        "required" : [ "itemId", "qtyOrdered", "unitPrice" ]
      },
      "CreatePoRequest" : {
        "type" : "object",
        "properties" : {
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CreatePoLine"
            },
            "minItems" : 1
          },
          "orderDate" : {
            "type" : "string",
            "format" : "date"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "lines", "orderDate", "partnerId" ]
      },
...
      "CreateReturnRequest" : {
        "type" : "object",
        "properties" : {
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "salesInvoiceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "stockLocationId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "salesInvoiceId", "stockLocationId" ]
      },
      "CreateSoLine" : {
        "type" : "object",
        "properties" : {
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "qtyOrdered" : {
            "type" : "string"
          },
          "unitPrice" : {
            "type" : "string"
          }
        },
        "required" : [ "itemId", "qtyOrdered", "unitPrice" ]
      },
      "CreateSoRequest" : {
        "type" : "object",
        "properties" : {
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CreateSoLine"
            },
            "minItems" : 1
          },
          "orderDate" : {
            "type" : "string",
            "format" : "date"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "lines", "orderDate", "partnerId" ]
      },
...
      "CustomerReturnLineView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lineCogs" : {
            "type" : "string"
          },
          "lineNet" : {
            "type" : "string"
          },
          "lineVat" : {
            "type" : "string"
          },
          "movementGroupId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "qty" : {
            "type" : "string"
          },
          "soLineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "stockJournalEntryId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "CustomerReturnResponse" : {
        "type" : "object",
        "properties" : {
          "cogsAmount" : {
            "type" : "string"
          },
          "creditNoteJournalEntryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "goodsAmount" : {
            "type" : "string"
          },
          "grossAmount" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CustomerReturnLineView"
            }
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "postedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "returnNumber" : {
            "type" : "string"
          },
          "salesInvoiceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "salesOrderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string"
          },
          "vatAmount" : {
            "type" : "string"
          }
        }
      },
...
      "InvoiceLine" : {
        "type" : "object",
        "properties" : {
          "qty" : {
            "type" : "string"
          },
          "soLineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "unitPrice" : {
            "type" : "string"
          }
        },
        "required" : [ "qty", "soLineId", "unitPrice" ]
      },
      "InvoiceLineView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lineCogs" : {
            "type" : "string"
          },
          "lineNet" : {
            "type" : "string"
          },
          "lineVat" : {
            "type" : "string"
          },
          "qty" : {
            "type" : "string"
          },
          "soLineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "unitPrice" : {
            "type" : "string"
          }
        }
      },
...
      "PoLineView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lineNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "qtyBilled" : {
            "type" : "string"
          },
          "qtyOrdered" : {
            "type" : "string"
          },
          "qtyReceived" : {
            "type" : "string"
          },
          "unitPrice" : {
            "type" : "string"
          }
        }
      },
...
      "PurchaseOrderResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PoLineView"
            }
          },
          "orderDate" : {
            "type" : "string",
            "format" : "date"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "poNumber" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
...
      "SalesInvoiceResponse" : {
        "type" : "object",
        "properties" : {
          "amountReceived" : {
            "type" : "string"
          },
          "cogsAmount" : {
            "type" : "string"
          },
          "goodsAmount" : {
            "type" : "string"
          },
          "grossAmount" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "invoiceNumber" : {
            "type" : "string"
          },
          "journalEntryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/InvoiceLineView"
            }
          },
          "openBalance" : {
            "type" : "string"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "postedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "postingDate" : {
            "type" : "string",
            "format" : "date"
          },
          "salesOrderId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "status" : {
            "type" : "string"
          },
          "vatAmount" : {
            "type" : "string"
          }
        }
      },
      "SalesOrderResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SoLineView"
            }
          },
          "orderDate" : {
            "type" : "string",
            "format" : "date"
          },
          "partnerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "soNumber" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
      "SoLineView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "itemId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lineNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "qtyInvoiced" : {
            "type" : "string"
          },
          "qtyOrdered" : {
            "type" : "string"
          },
          "qtyShipped" : {
            "type" : "string"
          },
          "unitPrice" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}